--- tests/models.py
@@ -1,72 +1,0 @@
-from enum import Enum
-from typing import Optional
-from datetime import datetime
-from pydantic import BaseModel
-import dotenv
-dotenv.load_dotenv(override=True)
-
-class TableOcr(str, Enum):
-    HTML = "HTML"
-    JSON = "JSON"
-    
-class SegmentationStrategy(str, Enum):
-    LayoutAnalysis = "LayoutAnalysis"
-    Page = "Page"
-
-
-class Model(str, Enum):
-    Fast = "Fast"
-    HighQuality = "HighQuality"
-
-
-class Status(str, Enum):
-    Starting = "Starting"
-    Processing = "Processing"
-    Succeeded = "Succeeded"
-    Failed = "Failed"
-    Canceled = "Canceled"
-
-
-class Configuration(BaseModel):
-    model: Model
-    target_chunk_length: int
-
-
-
-class OutputResponse(BaseModel):
-    chunks: list[dict]
-    extracted_json: Optional[dict] = None
-
-
-class TaskResponse(BaseModel):
-    task_id: str
-    status: Status
-    created_at: datetime
-    finished_at: Optional[datetime] = None
-    expires_at: Optional[datetime] = None
-    message: str
-    input_file_url: Optional[str] = None
-    output: Optional[OutputResponse] = None
-    task_url: Optional[str] = None
-    configuration: Configuration
-    file_name: Optional[str] = None
-    page_count: Optional[int] = None
-    pdf_url: Optional[str] = None
-
-    class Config:
-        json_encoders = {datetime: lambda v: v.isoformat()}
-
-
-class OcrStrategy(str, Enum):
-    Auto = "Auto"
-    Off = "Off"
-    All = "All"
-
-
-class UploadForm(BaseModel):
-    file: str
-    model: Model
-    target_chunk_length: Optional[int] = None
-    ocr_strategy: Optional[OcrStrategy] = OcrStrategy.Auto
-    json_schema: Optional[dict] = None
-    segmentation_strategy: Optional[SegmentationStrategy] = None
