--- tests/pyproject.toml
@@ -1,11 +1,0 @@
-[project]
-name = "tests"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.9"
-dependencies = [
-    "pydantic>=2.10.4",
-    "python-dotenv>=1.0.1",
-    "requests>=2.32.3",
-]
